Revert "feat: create foundation for fraudulent block production (#1992)"
This reverts commit 14c24b0c7c5a93e651539276c778b1a63d909d4b.
